requirements.txt
streamlit==1.32.2
google-generativeai==0.4.1
O365==2.0.32
imapclient==3.0.1
python-dotenv==1.0.1
